Changed hooks folder to api and renamed the Meeting folders
