+++ allData.json
@@ -1,0 +1,200 @@
+{
+  "meetings": [
+    {
+      "id": 1,
+      "room_id": 1,
+      "owner": 1,
+      "participants_id": [1, 2, 3],
+      "date": "2023-4-1",
+      "start_time": [12, 34],
+      "end_time": [13, 14],
+      "duration": 40
+    },
+    {
+      "id": 2,
+      "room_id": 1,
+      "owner": 3,
+      "participants_id": [1, 3, 4],
+      "date": "2023-4-1",
+      "start_time": [13, 40],
+      "end_time": [14, 10],
+      "duration": 30
+    },
+    {
+      "id": 3,
+      "room_id": 1,
+      "owner": 4,
+      "participants_id": [1, 2, 4],
+      "date": "2023-4-1",
+      "start_time": [12, 34],
+      "end_time": [12, 49],
+      "duration": 15
+    },
+    {
+      "id": 4,
+      "room_id": 2,
+      "owner": 3,
+      "participants_id": [3, 2, 5],
+      "date": "2023-4-2",
+      "start_time": [12, 34],
+      "end_time": [12, 49],
+      "duration": 15
+    },
+    {
+      "id": 5,
+      "room_id": 3,
+      "owner": 5,
+      "participants_id": [1, 4, 3, 2, 5],
+      "date": "2023-4-2",
+      "start_time": [12, 34],
+      "end_time": [12, 54],
+      "duration": 20
+    },
+    {
+      "id": 6,
+      "room_id": 3,
+      "owner": 4,
+      "participants_id": [1, 4, 3, 2, 5, 6, 7],
+      "date": "2023-4-3",
+      "start_time": [10, 34],
+      "end_time": [11, 4],
+      "duration": 30
+    },
+    {
+      "id": 7,
+      "room_id": 4,
+      "owner": 7,
+      "participants_id": [1, 4, 3, 2, 5, 6, 7],
+      "date": "2023-4-9",
+      "start_time": [9, 30],
+      "end_time": [10, 10],
+      "duration": 40
+    },
+    {
+      "id": 8,
+      "room_id": 4,
+      "owner": 1,
+      "participants_id": [1, 4, 3, 2, 5, 6, 7],
+      "date": "2023-4-9",
+      "start_time": [14, 30],
+      "end_time": [14, 45],
+      "duration": 15
+    },
+    {
+      "id": 9,
+      "room_id": 4,
+      "owner": 3,
+      "participants_id": [1, 4, 3, 2, 5, 6, 7],
+      "date": "2023-4-10",
+      "start_time": [15, 30],
+      "end_time": [15, 50],
+      "duration": 20
+    },
+    {
+      "id": 10,
+      "room_id": 5,
+      "owner": 4,
+      "participants_id": [1, 4, 3, 2, 5, 6],
+      "date": "2023-4-10",
+      "start_time": [14, 30],
+      "end_time": [15, 0],
+      "duration": 30
+    },
+    {
+      "id": 11,
+      "room_id": 5,
+      "owner": 6,
+      "participants_id": [1, 7, 3, 2, 5, 6],
+      "date": "2023-4-10",
+      "start_time": [16, 30],
+      "end_time": [17, 0],
+      "duration": 30
+    }
+  ],
+  "meetingRooms": [
+    {
+      "name": "Focus Room",
+      "id": 1,
+      "description": "This room is desgined for focused meetings and discussions",
+      "meetings_id": [1, 2, 3],
+      "capacity": 3,
+      "state": 0
+    },
+    {
+      "name": "Room for Growth",
+      "id": 2,
+      "description": "This room is desgined for growth meetings and discussions",
+      "meetings_id": [4],
+      "capacity": 3,
+      "state": 2
+    },
+    {
+      "name": "Collective IQ Room",
+      "id": 3,
+      "description": "This room is desgined for focused meetings and discussions",
+      "meetings_id": [5, 6],
+      "capacity": 10,
+      "state": 1
+    },
+    {
+      "name": "Everyones's Room",
+      "id": 4,
+      "description": "This room is desgined for focused meetings and discussions",
+      "meetings_id": [7, 8, 9],
+      "capacity": 12,
+      "state": 1
+    },
+    {
+      "name": "High Table Room",
+      "id": 5,
+      "description": "This room is desgined for focused meetings and discussions",
+      "meetings_id": [10, 11],
+      "capacity": 6,
+      "state": 2
+    }
+  ],
+  "participants": [
+    {
+      "name": "Alexandru Mariustiuc",
+      "id": 1,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Maria Popovici",
+      "id": 2,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Alexei Serghei",
+      "id": 3,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Sulei Marianovici",
+      "id": 4,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Alexa Amazon",
+      "id": 5,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Dora Barbat",
+      "id": 6,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    },
+    {
+      "name": "Flavia Mioc",
+      "id": 7,
+      "available": true,
+      "image": "https://w7.pngwing.com/pngs/205/731/png-transparent-default-avatar.png"
+    }
+  ]
+}
