--- src/Components/Admin/DeleteConfirmationModal/DeleteConfirmationModal.css
@@ -1,54 +1,0 @@
-.confirmation-modal {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: left;
-    background-color: #f8f9fa;
-    width: 600px;
-    height: 200px;
-    padding: 20px;
-    border: none;
-    border-radius: 5px;
-    position: fixed;
-    top: 40%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-  
-  .confirmation-modal h2 {
-    font-size: 24px;
-    margin-bottom: 20px;
-    font-weight: 500;
-  }
-  
-  .confirmation-modal p {
-    font-size: 16px;
-    margin-bottom: 15px;
-  }
-  
-  .confirmation-modal button {
-    margin-left: 10px;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-  }
-  
-  .confirmation-modal button.submit {
-    background-color: #7A7CFF;
-    color: #fff;
-  }
-  
-  .confirmation-modal button.cancel {
-    background: #FFFFFF;
-    color: #7A7CFF;
-    border: 1px solid #7A7CFF;
-    border-radius: 4px;
-  }
-
-  .confirmation-modal-buttons-layout{
-    margin-left: auto;
-    margin-bottom: 25px;
-  }
-
