--- src/CustomColors.ts
@@ -1,8 +1,0 @@
- const COLORS={
-       SUCCESS: "#40A368",
-        WARNING: "#BCA900",
-        ERROR: "#DD6764",
-        ADMINCOLOR: "#7A7CFF",
-        NOTAVAIBLE:"#EEEEEE",
-    }
-export default COLORS
