--- src/CustomSize.ts
@@ -1,11 +1,0 @@
-const SIZE={
-    HEIGHT_ADMIN_BUTTON:"50px",
-    HEIGHT_TABLE_BUTTON:"30px",
-    WIDTH_TABLE_BUTTON:"150px",
-    WIDTH_ROUNDED_TABLE_BUTTON:"200px",
-    HEIGHT_ROUNDED_TABLE_BUTTON:"30px",
-    HEIGT_CARD:"200px",
-    WIDTH_CARD:"400px",
-    BORDER_RADIUS_CARD:"10px"
-}
-export default SIZE
